add option to disable caching

+++ t
@@ -1,8 +1,5 @@
 #!/bin/bash
 
-EFF_TCACHE_DIR="${TCACHE_DIR:-$HOME/.tcmd-cache}"
-mkdir -p "$EFF_TCACHE_DIR"
-
 main() {
     check "$@"
     target="$(command ls -1tp -- *.{cpp,py,java,c,ml} 2>/dev/null | command sed 1q)"
@@ -13,7 +10,7 @@ main() {
     verbose_print "using base %s\n" "${base}"
     verbose_print "using extension %s\n" "${extension}"
 
-    if [ "$inputfile" = "" ]; then
+    if [ -z "$inputfile" ]; then
         verbose_print 'no input file given\n'
     elif [ ! -f "$inputfile" ]; then
         verbose_print '%s does not exist, resorting to stdin\n' "$inputfile"
@@ -22,12 +19,22 @@ main() {
         verbose_print "inputfile is %s\n" "${inputfile}"
     fi
 
+    if [ -z "${CACHE_OFF}" ]; then
+        verbose_print "using cache\n"
+
+        EFF_TCACHE_DIR="${TCACHE_DIR:-$HOME/.tcmd-cache}"
+        if [ ! -d "${EFF_TCACHE_DIR}" ]; then
+            error_print "cache directory %s does not exist; creating\n" "${EFF_TCACHE_DIR}"
+            mkdir -p "$EFF_TCACHE_DIR"
+        fi
+    fi
+
     run
 }
 
 check() {
     local OPTIND opt
-    while getopts ":vi:" opt; do
+    while getopts ":vci:" opt; do
         case "$opt" in
             i)
                 printf "input file %s\n" "$opt"
@@ -35,6 +42,9 @@ check() {
                 ;;
             v)
                 VERBOSE="verbose"
+                ;;
+            c)
+                CACHE_OFF="cache_off"
                 ;;
             \?)
                 print_help
@@ -51,13 +61,16 @@ run() {
             filehash="$(md5 -q ${target})"
             exppath="${EFF_TCACHE_DIR}/${filehash}"
             verbose_print "hash is %s\n" "$filehash"
-            if [ -e "${exppath}" ]; then
+            if [ -e "${exppath}" ] && [ -z "${CACHE_OFF}" ]; then
                 verbose_print "found cached @ %s\n" "$exppath"
                 cp "$exppath" ./a.out
             else
+                verbose_print "building\n"
                 gcc "$target" || exit $?
-                verbose_print "caching @ %s\n" "$exppath"
-                cp ./a.out "$exppath"
+                if [ -z "${CACHE_OFF}" ]; then
+                    verbose_print "caching @ %s\n" "$exppath"
+                    cp ./a.out "$exppath"
+                fi
             fi
 
             verbose_print "running executable\n"
@@ -72,13 +85,16 @@ run() {
             filehash="$(md5 -q ${target})"
             exppath="${EFF_TCACHE_DIR}/${filehash}"
             verbose_print "hash is %s\n" "$filehash"
-            if [ -e "${exppath}" ]; then
+            if [ -e "${exppath}" ] && [ -z "${CACHE_OFF}" ]; then
                 verbose_print "found cached @ %s\n" "$exppath"
                 cp "$exppath" ./a.out
             else
+                verbose_print "building\n"
                 g++-10 -DFEAST_LOCAL -std=c++11 "$target" || exit $?
-                verbose_print "caching @ %s\n" "$exppath"
-                cp ./a.out "$exppath"
+                if [ -z "${CACHE_OFF}" ]; then
+                    verbose_print "caching @ %s\n" "$exppath"
+                    cp ./a.out "$exppath"
+                fi
             fi
 
             verbose_print "running executable\n"
@@ -93,13 +109,17 @@ run() {
             filehash=$(md5sum "${target}" awk '{ print $1 }')
             exppath="${EFF_TCACHE_DIR}/${filehash}"
             verbose_print "hash is %s\n" "$filehash"
-            if [ -e "${exppath}" ]; then
+            if [ -e "${exppath}" ] && [ -z "${CACHE_OFF}" ]; then
                 verbose_print "found cached @ %s\n" "$exppath"
                 cp "$exppath" "${base}.class"
             else
+                verbose_print "building\n"
                 javac "$target" || exit $?
                 verbose_print "caching @ %s\n" "$exppath"
-                cp "${base}.class" "$exppath"
+                if [ -z "${CACHE_OFF}" ]; then
+                    verbose_print "caching @ %s\n" "$exppath"
+                    cp "${base}.class"  "$exppath"
+                fi
             fi
 
             verbose_print "running executable\n"
@@ -126,12 +146,13 @@ print_help() {
     printf -- 'usage:\n'
     printf -- '  -i <input file>\n'
     printf -- '  -v (verbose)\n'
+    printf -- '  -c do not use the cache\n'
 }
 
 verbose_print() {
     if [ -n "$VERBOSE" ]; then
         tput setaf 3
-        printf '| '
+        printf '(verbose) | '
         # reset tput style change
         tput sgr0
         printf -- "$@"
